feat(mcp-registry/beta): inline install + reworked detail Overview
- Extract the catalog install flow (install/add-personal/team/org, remote/
  local/no-auth dialogs, OAuth, enterprise-managed guard, secret handling)
  into a shared useCatalogInstall hook used by both the beta list and the
  beta detail page. Install now runs inline on the detail page instead of
  redirecting to the catalog. Reinstall/reauth stay in the list with their
  own dialog instances.
- Rework the detail Overview information architecture: uniform, non-interactive
  KPI tiles (Status / Credentials / Tools / Environment), a Tools capabilities
  panel (tool names + descriptions, Guardrails link, empty state), and a
  compact Details panel (type/transport, server URL or command, created,
  labels). Single-line scrollable command/URL.
